+++ crates/icydb-core/src/db/types/account.rs
@@ -8,6 +8,10 @@ use crate::{
     types::Account,
 };
 
+///
+/// Account
+///
+
 impl Filterable for Account {
     type Filter = TextFilterKind;
     type ListFilter = TextListFilterKind;