+++ crates/icydb-core/src/db/types/blob.rs
@@ -1,5 +1,9 @@
 use crate::{db::primitives::NoFilterKind, traits::Filterable, types::Blob};
 
+///
+/// Blob
+///
+
 impl Filterable for Blob {
     type Filter = NoFilterKind;
     type ListFilter = NoFilterKind;