+++ crates/icydb-core/src/db/types/decimal.rs
@@ -4,6 +4,10 @@ use crate::{
     types::Decimal,
 };
 
+///
+/// Decimal
+///
+
 impl Filterable for Decimal {
     type Filter = DecimalRangeFilterKind;
     type ListFilter = DecimalListFilterKind;