+++ crates/icydb-core/src/db/types/duration.rs
@@ -4,6 +4,10 @@ use crate::{
     types::Duration,
 };
 
+///
+/// Duration
+///
+
 impl Filterable for Duration {
     type Filter = Nat64RangeFilterKind;
     type ListFilter = Nat64ListFilterKind;