+++ crates/icydb-core/src/db/types/e8s.rs
@@ -4,6 +4,10 @@ use crate::{
     types::E8s,
 };
 
+///
+/// E8s
+///
+
 impl Filterable for E8s {
     type Filter = Nat64RangeFilterKind;
     type ListFilter = Nat64ListFilterKind;